Absorb `_utils.py` into `equal_metadata.py`
