+++ python/src/equistore/operations/lstsq.py
@@ -5,7 +5,7 @@ import numpy as np
 from ..block import TensorBlock
 from ..tensor import TensorMap
 from . import _dispatch
-from ._utils import _check_maps, _check_same_gradients
+from .equal_metadata import _check_maps, _check_same_gradients
 
 
 def lstsq(X: TensorMap, Y: TensorMap, rcond, driver=None) -> TensorMap:
