+++ python/src/equistore/operations/multiply.py
@@ -5,7 +5,7 @@ import numpy as np
 from ..block import TensorBlock
 from ..tensor import TensorMap
 from . import _dispatch
-from ._utils import _check_blocks, _check_maps, _check_same_gradients
+from .equal_metadata import _check_blocks, _check_maps, _check_same_gradients
 
 
 def multiply(A: TensorMap, B: Union[float, TensorMap]) -> TensorMap:
